bonne avancé

+++ src/Controller/CaveController.php
@@ -26,13 +26,14 @@ class CaveController extends AbstractController
     }
 
     #[Route('/new', name: 'app_cave_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, CaveRepository $caveRepository): Response
+    public function new(Request $request, CaveRepository $caveRepository, UserInterface $user): Response
     {
         $cave = new Cave();
         $form = $this->createForm(CaveType::class, $cave);
         $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
+        var_dump($cave->getUtilistaeurId());
+        var_dump($user);
+        if ($form->isSubmitted() && $form->isValid() && $cave->getUtilistaeurId() === $user) {
             $caveRepository->save($cave, true);
 
             return $this->redirectToRoute('app_cave_index', [], Response::HTTP_SEE_OTHER);
@@ -45,8 +46,13 @@ class CaveController extends AbstractController
     }
 
     #[Route('/{id}/edit', name: 'app_cave_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Cave $cave, CaveRepository $caveRepository): Response
+    public function edit(Request $request, Cave $cave, CaveRepository $caveRepository, UserInterface $user): Response
     {
+        // Vérifier si l'utilisateur associé à la cave correspond à l'utilisateur actuel
+        if ($cave->getUtilistaeurId() !== $user) {
+            // Rediriger l'utilisateur vers une autre page
+            return $this->redirectToRoute('app_cave_index');
+        }
         $form = $this->createForm(CaveType::class, $cave);
         $form->handleRequest($request);
 
@@ -63,9 +69,9 @@ class CaveController extends AbstractController
     }
 
     #[Route('/{id}', name: 'app_cave_delete', methods: ['POST'])]
-    public function delete(Request $request, Cave $cave, CaveRepository $caveRepository): Response
+    public function delete(Request $request, Cave $cave, CaveRepository $caveRepository, UserInterface $user): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$cave->getId(), $request->request->get('_token'))) {
+        if ($this->isCsrfTokenValid('delete'.$cave->getId(), $request->request->get('_token')) && $cave->getUtilistaeurId() === $user) {
             $caveRepository->remove($cave, true);
         }
